requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
django-filter>=23.2
psycopg2-binary>=2.9
openpyxl>=3.1
pandas>=2.0
python-dotenv>=1.0
gunicorn>=21.2
